allow pelecan to properly reconnect after server restarts
modify implementation details of ClientManager's start connection method such that the new thread is spawned there instead of in the extrenal start_connection function. This allows the clients field to be changed to an Arc<Mutex<_>> which means that clients can be removed after the connection closes if it is detected that the client has disconnected. This means a new client connection will be established upon a reconnect attempt.